+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog for autoNumeric
 
+### 4.2.1
++ Fix issue #553 Missing support for very small or very big numbers displayed by Javascript as scientific numbers
+  + This adds support for using scientific notation for setting the input value (ie. `aNInput.set('6.1349392e-13');`, `<input value="7342.561e40">`)
+
 ### 4.2.0
 + Fix issue #535 Prevent entering any decimal character when only positive numbers are accepted
 + Change how the decimal character can be entered: